Added skeleton of main tex file. SEE NOTES BELOW!
I just quickly made a skeleton of the main tex file. I'm trying as best
I can to according to Kristian keep format and text separated. So that's
new, and fun!

Add more packages if you want. I just didn't care enough as to add more
of them right now.

And for this report, how about not adding the norwegian paragraph
package? I think it looks nice without for a change!